V-6.1
JavaScript function change_display added to alert.js to allow instant submit when clicking on one of radio buttons that describe the result layout.
This function was added so that the user can resubmit the query for a different display format.

A show/hide for Advanced Settings was added.
	Stat	ON/OFF	(option hidden by default)
	Aggregation ON/OFF	(option hidden by default)

search_test_3.html created to test this layout.